requirements.txt
emoji-country-flag
fastapi
jinja2
pycountry
pygit2
python-dotenv
python-slugify
requests
tqdm
uvicorn
